LayoutTests/http/tests/websocket/tests/hybi/handshake-ok-with-http-version-beyond-1_1.html

<!DOCTYPE html>
<html lang="en">
<head>
    <script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
    description("Test whether WebSocket handshake is OK with if server sends status line with http version beyond 1.1.");

    window.jsTestIsAsync = true;

    var url = "ws://localhost:8880/websocket/tests/hybi/handshake-ok-with-http-version-beyond-1_1";
    var handshake_success = false;
    var ws = new WebSocket(url);
    var closeEvent;

    function endTest() {
        clearTimeout(timeoutID);
        shouldBeTrue("handshake_success");
        finishJSTest();
    }
    ws.onopen = function() {
        handshake_success = true;
        ws.close();
    };

    ws.onclose = function () {
        endTest();
    };
    
    var timeoutID = setTimeout("endTest()", 2000);
</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
