LayoutTests/http/tests/websocket/tests/hybi/handshake-ok-with-http-version-beyond-1_1-expected.txt

Test whether WebSocket handshake is OK with if server sends status line with http version beyond 1.1.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS handshake_success is true
PASS successfullyParsed is true

TEST COMPLETE
